+++ src/db-luna/MojDbBerkeleyQuery.cpp
@@ -84,7 +84,7 @@ MojErr MojDbBerkeleyQuery::getById(const MojObject& id, MojDbStorageItem*& itemO
 		MojErrCheck(err);
 		if (!foundOut) {
 			char s[1024];
-			int size = primaryKey.size();
+			int size = (int)primaryKey.size();
 			(void) MojByteArrayToHex(primaryKey.data(), size, s); 
 			MojLogInfo(MojDbBerkeleyEngine::s_log, _T("bdbq_byId_warnindex: KeySize: %d; %s ;id: %s \n"), 
 								 size, s, primaryKey.data()+1);
